pyproject.toml
[project]
name = "ragfirsttest"
version = "0"
description = ""
authors = [
    {name = "Arseny"}
]
readme = "README.md"
requires-python = "^3.12,<3.15"
dependencies = [
    "numpy (>=2.3.4,<3.0.0)",
    "langchain (>=1.0.1,<2.0.0)",
    "langchain-community (>=0.4,<0.5)",
    "pandas (>=2.3.3,<3.0.0)",
    "torch (>=2.9.0,<3.0.0)",
    "transformers (>=4.57.1,<5.0.0)",
    "opensearch-py (>=3.0.0,<4.0.0)",
    "pymupdf (>=1.26.5,<2.0.0)",
    "langchain-text-splitters (>=1.0.0,<2.0.0)",
    "langchain-ollama (>=1.0.0,<2.0.0)",
    "mlflow (>=3.5.1,<4.0.0)",
    "openai (>=2.6.1,<3.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
